Create new widget FileUploadUIAR
Getting close to operational - just a progress backup. This just prior
to use active form js if possible to update form input errors.
actionUpdate now working well with jquery-file-upload and appending a
redirect parameter in the returned json after upload initiated by
jquery file upload. This is redirecting on success and have a place in
fileuploadfinished callback to deal with errors that are returned in
the event of an error. If neither redirect nor form errors returned
then file errors must have occurred. Have altered uploadHandler in
non-obtrusive manner to work with this. May request blue imp to apply
changes to UploadHandler and to add jquery-file-upload to packagist or
bower if possible - not sure if php files can be added to bower yet as
it claims to be a front end package manager an d no backend